Refactor: Extract obj functions to external files
