+++ src/objectFunctions/updateObj.ts
@@ -1,0 +1,6 @@
+import { cloneObject } from './cloneObject';
+
+/** @desc Update input Object */
+export const updateObj = <T>(obj: T, fieldsToUpdate: Partial<T>): T => {
+  return { ...cloneObject(obj), ...cloneObject(fieldsToUpdate) };
+};
